[FIX] core/routing/Target: _updateOptions should not overwrite _name
It turned out that if subroutes is used mixed with the target
configuration, _updateOptions is called more than once for
targets created subroutes. Once in the constructor, once in
the Route which uses the implicitly created Target. To avoid
side effects, the fix has been placed in _updateOptions instead
of removing the actual _updateOptions calls.

Change-Id: Ie56b71683fd148d9744fd03475e588368fc011e4
BCP: 1870335099
CR-Id: 002075125800003111532018