[FIX]sap.ui.rta - Actions no longer available for wrong aggregations
When checking for actions on aggregations we were not considering the
aggregation where the element requesting the check was located. This fix
includes this extra check ensuring that the action will only be
available on the elements of the defined aggregation.

CR-Id: 002075125800003145122018
BCP: 1870346324

Change-Id: Ia0bafa5e93276628c7a67c6da9f00ef733e9105e
Signed-off-by: Vitor Seifert <vitor.eduardo.seifert.bazzo@sap.com>
(cherry picked from commit e97a17d737235ed6a9f9c162a6df9bf41502a0c9)