[FIX] Image: Would not fire error if the source is invalid
The Problem: Image would not fire error if the source is empty
or not existing.

The Solution: Image is firing error when the source is empty or
not existing, even if we have no alt text available.

BCP: 1880526262
(cherry picked from commit 4feef23aea2970991d4855d4aa5f1cf725a16ae0)

Change-Id: Idaa5e68d751a6017b5e3225d063b0e9a3a9531ea
CR-Id: 002075125800003171122018