[FIX] sap.m.StepInput: correct behavior when mousewheel is used
Previously the native behavior of the input is used and when
a mousewheel is used it increses and decreses the value.
However, in IE and Edge this is not supported.

Furthermore since we do not set max and min attributes to the
input, the validation did not work in Chrome too.

Now mousewheel events are handled properly

Change-Id: I0b10f771be0083bcd534e3f56c90da93149e7bba
BCP: 1870334890