[FIX] sap.ui.unified.FileUploader: correct width of the input when invalidated
FileUploader uses custom setters for some of its properties in order to
propagate the value to the internally create input. However, some of
these properties do not use custom setter in the input and this causes
invalidation of the input but does not trigger re-rerender of the
FileUploader. In these cases the logic for width calculation of the input
is not triggerd.

Currently two properties causes input invalidation without the re-rerender
of the FileUploader: setEnabled and setPlaceholder.

In these cases an event delegate to the onAfterRendering of the input is
attached and the logic for width calculation is called from there.

BCP: 1870362150

Change-Id: Iae8640159389fae0afa0d58dd93ddfde06234b73
CR-Id: 002075125800003342832018