[FEATURE][INTERNAL] flex sends a CSP compliant server request for extensions
With this change js code extensions as well as xml fragments are no longer part of the initial request but is requested in another one which is of the type application/javascript.
The second request is only manadatory for legacy coding. Further steps in the best practice build of applications will avoid this call.

Change-Id: Ia33e1bdc6435d741015fb55c9a238942cd63c1ab