[FIX] format/DateFormat: Multiple instances shouldn't share fallback formats
When multiple DateFormat instances of different types (interval, non-interval etc.)
are created the fallback formats are merged in a static date info object.

This lead to the issue, that the wrong fallback format was taken.
To solve this issue, the fallback formats are created separately for every single
date format instance.

BCP: 1870290668
Change-Id: I9496f5e17f82fdfc719762bbb69b1c5509e8442a