[FIX] sap.m.StepInput: correct valueState when setValue is called
Order of calling property setters is not guaranteed and it could vary
on every control initialization and browsers.

_verifyValue method is called from min and max setters to ensure that
a proper valueState is set to the control (because changing the min or
max can put the value outside of the valid range). However, changing
the value could also put the value outside the valid range and should
also change the valueState.

The described not guaranteed calling of the setters could be easily
recreated event with some configurations:
create a new sap.m.StepInput with some min and max and a value outside
of the valid range. This will put the control in Error state.
After that if a setValue method is called with a value that fits in the
range the control won't become with valueState none.

BCP: 1870329912

Change-Id: Ic4ca589f897d9777643b8e8f8cc58a1434be80fc