[FEATURE]Image: densityAware default is changed to "false"
Before 1.60, the default value of the "densityAware" property of sap.m.Image was true,
which brought redundant network requests for applications that used the default
but did not provide density perfect image versions on server-side.

We got permission to change the default to false since 1.60, so that apps enable this property
only if they also provide the corresponding image versions for high-density devices.

Change-Id: I74666b279f64ca800aad16e42211b2211f5290b2
JIRA: BGSOFUIPIRIN-3126