[FIX] Fragment: remove new property names before calling legacy API
The new asynchronous Fragment.load() API introduced new property names
in its options object for the fragment's name, its definition and
controller (using the chance of a new API to use shorter and more
precise names).

As the old and the new API currently share the same implementation,
those properties are internally normalized to the old names. But they
had not been removed from the options object and therefore caused
assertion failures in the old implementation.

This has been fixed now.

Fixes: https://github.com/SAP/openui5/issues/2193
Change-Id: I64556ca4bc719fd1a8c396e4ccea1de52535081f
BCP: n/a