[INTERNAL] ObjectPage: Move focus to subsection in IconTabBar mode fixed
BCP: 1870329491

Problem: Upon clicking on a submenu button leading to subsection - the
focus is not moved to the subsection only in IconTabBar mode.

Solution: Allow focus to move in IconTabBar mode for subsections only,
since it's pointless for sections(every tab has one section only).

Actually this was never implemented, so I set it as internal here.

Change-Id: I7b1712c47ab38ffcf069d7b8884c8fb4d2c8bc2d
CR-Id: 002075125800003773642018