[FIX] AlignedFlowLayout: Accessiblity issue fixed
The issue:
In some scenarios, when accessing the first focusable (HTML element/
control) rendered inside an AlignedFlowLayout's item via the keyboard,
all other focusable (HTML element/controls) rendered inside the same
item are also announced by the screen reader (JAWS 18).
However, the expected behavior in this case is that only the focusable
(HTML element/control) should be announced by the screen reader.

The root cause of the issue:
The issue is caused by the screen reader if the following conditions
apply:

- The ARIA application mode is set
(ARIA role="application" is set on the body element of a HTML document,
which is the default accessibility mode in UI5).

- An unordered list is used to render the content.

The solution:
Replace the usages of the unordered list with div elements.

Change-Id: I082be62ca4c0d8da38e4814be6783ee9e103d825
BCP: 1870364846