[INTERNAL] ObjectPageLayout: Support for subSectionPreload event
Background:
-----------
The sap.uxap.ObjectPageLayout control supports "preLoading" and "lazyLoading"
of (sub)sections.

1) "preLoad": *Before* the page is rendered, the data for the first X sections
is loaded early (where X depends on the platform)

2) "lazyLoading": *After* the page is rendered, the data for the (sub)Sections
that are visible in the viewport is loaded.

For the (sub)sections that are eligible to lazyLoading,
the sap.uxap.ObjectPageLayout also fires the "subSectionEnteredViewPort" event
(to notify apps that want to do action upon lazyLoading).

Problem:
--------
The apps want to do action upon "preLoad" of (sub)sections
(and not only on "lazyLoading"), but no event
(corresponding to "subSectionEnteredViewPort") is fired on "preLoad".

Solution:
---------
Fire "subSectionPreload" event on "preLoad" of subsections.

Change-Id: I132313e4da8188e553f8d1a08bd15551413d2f06
BCP: 1880479448