[FIX] ui.layout.Grid: Control visibility properties behavior fixed
BCP: 1870416296

Problem: There is some old logic, causing controls to be shown on given
screen sizes, even when their corresponding visible property is set to
false. Also there is a mixture of visible and hidden classes set which
is confusing.

Solution: Use straight-forward approach to set only a hidden class for
the given screen size if the corresponding control's visible property is
set to false

Change-Id: I4e5737e041667c30f7f3e0fd8509c286bf1f47d0
CR-Id: 002075125800003737482018