[INTERNAL][FIX] Model all library dependencies also as AMD dependencies
When multiple libraries are loaded asynchronously, all request are
async, but the order in which the library.js modules are executed,
is only determined by their modeled relative dependencies.

When a dependency from one lib to another is __not__ modeled as part of
the AMD signature, it might happen that the dependent lib is executed
before the lib that it depends on. During the initLibrary call, this is
detected and compensated by the Core with a sync request for the
dependency. Due to the already pending async request for the lib, this
causes an async/sync conflict, an additional request (async/sync
conflicts always create additional requests) and the sync request is not
compatible with content security policies that forbid 'eval'.

Once 'manifest first' will be implemented for libraries, the Core can
fix this internally. Until then, adding the dependencies is an easy and
straight forward fix.

Change-Id: I12e3725000ec94bfb3bdec44b002c961bc1a76cc