[FIX] m/TextArea: do not trigger FieldGroupValidation on enter
Introduced a marker on the keyup event to enable the UIArea to
prevent the FieldGroupValidation. As the handling of the Enter
key is special in TextArea and the UIArea should not know about
certain Controls and does not know about the context of the event
this loose coupling needs to be introduced.

Change-Id: Ibbafc89c768f5e8519b6d080e13a74d28d06359a
BCP: 1870406549
(cherry picked from commit 10a063692ff7f04e723ba7697ad1aec2baad4ef8)
CR-Id: 002075125800004115302018