[FIX]sap.ui.fl - Fixed switching variants with stashed controls
We can't save applied changes on the customData of
sap.ui.core._StashedControl because after unstash these controls no
longer exist. Instead we assume that a stash operation on these controls
is always revertible and verify the customData of the unstashed control
(e.g. the object page section).

Note: a test for this fix will be included in a separate change

Change-Id: Ia51dbd8da09ab1ac1446e7ecc1ce43044e63ec6e
Signed-off-by: Vitor Seifert <vitor.eduardo.seifert.bazzo@sap.com>