[INTERNAL] SDK: Stabilized old format topic and API URL's redirect
The problem:
Old style topic and API Reference URL's sometimes fail to redirect
causing JavaScript exception to be thrown.

The root cause:
Router pattern match method that was provided created a second View and
a second controller using a non-standard approach for this URL which
had different context than controller created from the router.
This caused a JavaScript exception to be thrown and navigation fails.

The solution:
Use standard Router.navTo method to navigate for this scenarios.

Change-Id: Ia78d5af053ea173be16d06ceb28db7854461c6d4
BCP: 1870382426