[FIX] SegmentedButton: Fire event for OverflowToolbar when width is changed
Problem: When SegmentedButton is placed in OverflowToolbar and it
internally changes its DOM element width, it does not notify the
OverflowToolbar, which as already cached the old width value. This leads
to cutting of part of the OverflowToolbar's content.

Solution: Implement IOverflowToolbarContent interface in
SegmentedButton. Create an OverflowToolbar config with all needed
options. Fire event each time when DOM element width of
SegmentedButton is internally changed, using _onInvalidationEventFired
handler of OverflowToolbar, which is attached to all invalidation events
of controls, implementing IOverflowToolbarContent. This way,
OverflowToolbar know when it needs to newly cache the sizes of its
containing elements.

Change-Id: I2dfc58699241939bf87f3c369d5b9faa795b3f51
BCP: 1870423080