[INTERNAL] SDK: Terms of Use menu target is calculated correctly
The problem:
Sometimes Terms of Use is pointing to corporate page instead of local
TermsOfUse.txt even for OpenUI5 distributions.

The root cause:
OpenUI5 runtime is calculated based on a model that is filled by a async
operation but the object key value is filled in a sync way. That results
in this.getModel("versionData").getProperty("/isOpenUI5") to return
undefined and thus setting the wrong target for the menu.

The solution:
Make by default the corporate page the default target and update it
in a async manner when the async versionData is available.

Change-Id: I85776f2ab3a9bd52cf6025bd99ac71f3fbd020f4