[INTERNAL][FIX] Table: Prevent selection in SelectionMode=None
If selectionMode is "None", selection is still possible via some
parts of the selection API. But the table will not immediately update
the UI, because _updateSelection simply returns in this mode. In the
renderer there are no checks like this though, so the table will show
selected rows after re-rendering.

If the SelectionMode is "None", selection of rows should not be possible
at all.

Change-Id: Ic3808cb1976cfa7c76516995b24f1b367a97b807
BCP: 1870340068