[FIX] sap.f.DynamicPage: Preserved scroll position on show/hide
Issue
Hiding the DynamicPage via display:none (NavContainer use case)
and showing the page again causes lost of previously reached scroll position.

Root cause
When the DynamicPage gets hidden, the custom scroll bar is updated
via oScrollbar.setContentSize(iSize), where in this case iSize is 0,
because the page has no size.
And, this causes the scrollTop of the Scrollbar`s internal div element to become 0 as well.

Solution
Skip updating the contentSize of the ScrollBar in these curcumstences.
This is also kind of optimization as the update won`t change anything
as the whole page is not visible.
When it gets visible again the update will take effect as currently does.

Change-Id: I7c11bd76ea46d852db80a7d2255b9d5fa965b24f