[FIX] sap.ui.unified.CalendarTimeInterval: navigates to the correct date
Previously when navigating trough the calendar to the last date of the
previous month, the CalendarTimeInterval was selecting wrong date
(in most scenarios - the first date of the previous month).

Now it selects the correct date - the last one of the previous month.

BCP: 0020751294 0000518213 2018

Change-Id: I08acdceadad92db3e8080df24627f73265bd93a2
(cherry picked from commit 6836691409bee19f24a0cb159d6871aad1e32c67)
CR-Id: 002075125800004198692018