[INTERNAL] ui5loader-autoconfig: add shims for more non-UI5 modules
Having the right shims configured in time becomes more important with
async loading being available.

Ideally, the shims for the sap.viz library should not be part of the
ui5loader-autoconfig.js. But as long as the descriptor for libraries is
not evaluated before loading the library (AKA 'manifest-first for
libs'), there's no other suitable point in time to apply the shims early
enough.

Change-Id: Ideac97fa88f4e2b432069bad58f46df9ac4f9e17