[INTERNAL] Infra: support array syntax for nested param documentation
The JSDoc3 toolkit supports documenting the properties of method and
event parameters by providing additional @param tags with qualified
names (paramName.property, also nested). When a parameter has an array
type, the qualified name can be composed from the parameter name, a pair
of square brackets (as array indicator), a dot and the property name:

  @param {object[]} myArrayParam
  @param {string} myArrayParam[].name
  @param {string} myArrayParam[].type

UI5's JSDoc template did not support this array type syntax. This has
been added as an option, but only when the parameter type is of an array
type ("array", "something[]" or "Array.<something>").

The representation in the api.json has not been modified. The properties
of the parameter type are specified in a nested "parameterProperties"
object. The only indicator for the array type use case is the type
property of the parameter itself.

Documenting the structure of array types is also supported on deeper
levels of structuring (if a property of a property of a parameter has
an array type).

Change-Id: Ic85b312d26db0d45f77757ab63bc931d310fbae6