[FIX] sap.uxap.ObjectPageSubSection: Changed header rendering
Background
The SubSection header is not rendered when
(1) the title property is empty string
(2) the title property value contains spaces only.

Issue
The described behavior (oprimization) causes issue in RTA,
when the user tries to rename SubSection title, specifically when delete
the whole title. There is no way to rename it back
as the SubSection header will not be rendered.

Solution
Render the SubSection header if the title contains spaces only.
Still, for an empty string, the header will not be rendered.
But, this case will be handled in RTA layer - when the user deletes the whole title, the title will be set to spaces only.

BCP: 1870437348
Change-Id: I10ea66492273580f9cc8a38d11e351cd9603cba3