[FIX] sap.uxap.AnchorBar: Fixed tab focusable state
Background
All buttons we are using for tabs are focusable by default.
We change their tabindex through the _setAnchorButtonsTabFocusValues method,
that apply tabindex 0 to the selected tab and -1 to all the rest.

Issue
Upon navigating to ObjectPage (via NavContainer) the first tab is getting the focus,
although the second tab will be the selected one.

Root cause
The _setAnchorButtonsTabFocusValues is executed is too late,
the first tab already has taken the focus.

Solution
Call _setAnchorButtonsTabFocusValues early as possible,
even when the selected button is not known yet, e.g all tabs will have tabindex -1

Change-Id: I786eef804cbb15ef6de4acde854b48f1c23231fc
BCP: 1870391330