[FEATURE] Control.extend: allow to specify ready-made renderers
When controls declare a dependency to their renderers, it seems
appealing to provide the renderer object in the class definition
(property 'renderer' of the class info object).

But so far, any given object was automatically enhanced with the
renderer of the base class, something that ready-made renderers have
already done.

With this change, most ready-made renderers are recognized and used
without further action. Additionally, ElementMetadata now caches the
renderer for further usage.

Change-Id: I9a4daf1048342445bce78b5bde185af489a20b3b