[INTERNAL][FIX] sap.ui.rta No add if parent misses stable ID
If there is a relevant container, different from the parent, we missed
to check for the stable ID of the parent control as well as the relevant
container.
The same is true if the hidden element has a different parent than the
sibiling, next to which the element should be added.

With this change it is fixed.

Change-Id: Ia7107047bacfd9146788a90d53a4209bf92b1c83
BCP: 1880561481