[INTERNAL] sap.m.SuggestionsPopover: New popover extracted from sap.m.Input
This change is a refactoring of the sap.m.Input that extracts the
suggestions functionality into a new file. The is the first of several
changes. It moves the methods that are related to the suggestions popover
into the new file without changing the names for for easy comparison.
Inside the methods only the references are changed. Another major change
is delegating the key events on the input to the new control.

Changes:
- patchset 1: initial change
- patchset 2: internal variable used by other controls restored
- patchset 3: references in _getShowMoreButton() corrected
- patchset 4: internal variable _oSuggestionTable also restored
- patchset 5: var _oSuggestionTable initialized where table created
- patchset 6: references corrected and a condition simplified
- patchset 7: references corrected

Change-Id: I364539c8790bbd32f8ef9c840273274eedafb831