[INTERNAL] sap.ui.core.Hyphenation: Unstable qUnit was removed
The problem here was with different support of "ru" language
on different platforms. For example on windows 10 IE11 and Edge
combi device the language is not supported but on windows 10
which is not combi the "ru" is supported. We have removed the
test because this particular matrix check may lead to falling test
in later time.

Change-Id: I09f75a3d56bb6daa48c6b07c40d7322c5bf58078