[FIX] Core: detect bootstrap script even when query params are present
Karma test runner adds a hash to the URL for cache busting. When this is
done OpenUI5 doesn't detect the sap-ui-core.js file anymore and tests
cannot be executed. This is fixed with this change.
See discussion in karma-openui5 repository.

Additionally, all the regexp patterns used to detected the bootstrap
script are aligned with each other.

Fixes: https://github.com/SAP/karma-openui5/issues/3
Closes: https://github.com/SAP/openui5/pull/2247
Change-Id: I555c4711a0410f22ec1ec08b13415caa271005ab
BCP: 1880623957