[FEATURE] ui.unified.Month: Keyboard interval selection aligned
Keyboard navigation in single interval selection mode is now
aligned with its mouse counterpart. This means that if a user
has selected a range's startingDate, but not endDate yet, then
any date between startingDate and the currently focused date
will indicate that it will be part of the selected interval.

Change-Id: I16994f8befcd51273776b7e660c2aee80788f5b0
JIRA: BGSOFUIBALKAN-4243