[FIX] performance/trace/FESR: deal with unsupported window.Performance API
As the Fiori Client with a WKWebView does not support the proper Performance
interface as described in https://www.w3.org/TR/performance-timeline/#extensions-to-the-performance-interface it is not possible to retrieve useful performance date.
Hence FESR is not activatable on such clients via script meta-tag or URL
parameter.
Furthermore the availability check has been reintroduced on the legacy API.

BCP: 1870446987
CR-Id: 002075125800004859512018
Change-Id: Ieb2955a1e63c7bb475c92f02bd8fffc5f0178009
(cherry picked from commit 08c40843ca84d19de79a402b311376690775eda6)