[INTERNAL] sap.ui.unified.ColorPicker: Now the grid is removed
JIRA: BGSOFUIBALKAN-4471

Now the unified.ColorPicker does't use Grid or Vertical/Horizontal layout.
The button in Default display mode follows the latest visual design
spec - before that it was cut at half.
Issues fixed:
- Now colorString property is in sync with the right input type.
Before the fix the getter was returning every time only the rgb value in
Large mode even if the HSL was selected via the radio buttons.
- Now when on phone, the comparison field for the new color is accordingly
filled with the corresponding image when the alpha slider is changed.

Change-Id: I7879c3189c8055e958cf4e277341f0725deb1321