[INTERNAL] Core: improve test finder app
The createSuite.js module of the generic test starter now adds the
whole config object for a test to the created suite (as second
parameter of addTestPage).

Normal test execution in voters and in the nightly build ignores this
second parameter. But the test finder app uses it to make more config
properties searchable, esp. the qunit and sinon version as well as the
set of executed modules.

Change-Id: I6ca56799d302c73aa1d6254430f22775f7d184d3