[FIX] m.DateTimePicker: maxDate being increased fixed
maxDate was being increased with one day when checking if
the initial focused date was between minDate and maxDate.
This way if maxDate was 'yesterday', but the date to be focused
was 'today', then an error would've been thrown on open, since
'today' would've appeared as valid date in this check, while it
was actually invalid.

Change-Id: I133a86cfd440c7da69a8da7d1ebf83c57721782e
BCP: 1880654028
Fixes: https://github.com/SAP/openui5/issues/2261