[FIX] sap.m.SegmentedButton: Width changed is not fired when in Popover
Problem: When SegmentedButton moves in the Associative Popover of
OverflowToolbar it changes its width and fires _containerWidthChanged
event, which invalidates the OverflowToolbar. This leads to closing the
Popover, which is not opened yet. The result is error in the Console and
the Popover cannot be opened.

Solution: SegmentedButton does not fire _containerWidthChanged event
when it is placed in the Associative Popover of OverflowToolbar. The
event is only needed to be fired when SegmentedButton is placed in
the Toolbar and one of its buttons loads its icon and updates its size.
In this case OverflowToolbar should be informed to adjust its cached
controls' widths.
QUnits are also updated in a way, which better reflects the real case
situations.

Change-Id: Ie095f22a52da76f76ef1b5b3f141638aa400fea0
CR-Id: 002075125800004990292018
BCP: 1880653248
(cherry picked from commit fe1eccbd281b7bb3d2efd969eddd47c8f60b515b)