[FIX] sap.uxap.ObjectPageLayout: Empty SubSection title disabled in RTA
Giving an empty title of SubSection in ObjectPageLayout is now disabled
in RTA scenario - now an Error in the Console is logged and the action
is not saved. This behavior fixes two problems:
- Giving an empty title of a SubSection results in not being able to set
a new title as the control is getting invalidated and the empty
placeholder is not rendered (now the control is not invalidated and the
empty value is not saved, so the placeholder is still there for editting
if page is not refreshed and if it is refreshed - the inital title shows
up. Now we can savely revert the fix in ObjectPageSubSectionRenderer,
which has removed trimming of the title's value when checking whether
the SubSection has title).
- If setting an empty title of an ObjectPageSubSection is allowed,
giving a non-empty title to SubSection, then - giving an empty title,
and after that - giving again a non-empty title results in returning
the original title of the SubSection, not the firstly changed and valid
one (now firstly changed and valid one is returned).

Change-Id: Ie8dcdbc5851f55fcc86c5aabbfb6632624fc4d7f
BCP: 1880499390