[FIX] sap.m.Select: Fixed focus outline offset for contrast themes
Problem: In high-contrast themes when the Select has focus,
the outline offset was incorrect and the outline was rendered
right next to the Select's border. The problem was caused by
incorrect calculations for the outline-offset as the outline is
0.125 rem.

Fix:
Fixed the outline-offset by compensating for the thicker outline.

Change-Id: I1b36e38db0d472199f24ed358cb07a8734b57a90
BCP:1880654497