[INTERNAL] Core: remove obsolete 3rdparty readme, fix license link
The overview page 'sap/ui/thirdparty/readme.html', which lists all
3rdparty components that are included in a UI5 version, hasn't been
referenced by the SDKs for a long time. Instead, the 'About' dialogs in
the SDKs either refer to the LICENSE.txt (OpenUI5 SDK) or to a view that
collects and displays the same information (SAPUI5 SDK).

Instead of making readme.html CSP compliant, it is removed.

Additionally, a typo in the file name of the lodash license copy has
been fixed so that incoming links work again (e.g. from 'About' dialog)

Change-Id: I1f2a246b71ed4497bad63c3c890a75561d36f338