[INTERNAL] ui5loader: reuse preload cache in case of async/sync conflict
1. AMD module (simple: unnamed or name matches requested module)

When the requested module was contained in the preload cache and when it
uses sap.ui.define, it won't complete synchronously. But for the same
reason it doesn't immediately clear the content of the preload cache.
Only the module status prevented a reuse of the preload cache content
when a conflicting sync request occurred (no longer PRELOADED,
but EXECUTING). This change detects this case and restores the PRELOADED
state.

2. AMD module (alias: no module definition resolves with the module
name)
Same as 1.

3. jQuery.sap.declare module (without cycle)
As such modules have no async contract, the state of the module will be
READY after the execModule() call. There shouldn't be an async/sync
conflict at all.

4. jQuery.sap.declare module (with cycle)
Re-execution cannot be prevented as the sync/async conflicts can only
be detected in the middle of the first execution

5. global script (or 3rdparty using UMD + shim)
This kind of script also has no async contract when executed by the
ui5loader, so there should be no conflict as well.

Change-Id: Ie25c62ebad28a486de5c900c99756f63753f20df