[FIX] sap.uxap.ObjectPageHeader: Corrected layout calculation
Problem:
The ObjectPageHeader was resized during its layout calculation
(as the layout calculation always removes the "sapOPHSubtitleBlock"
class to restore the default state)

The resizing of the header causes an *extra scroll event*
in the ObjectPageLayout in the case when the page has:
- snapped header
- no extra scrollable content becides the hidden content
The provided qunit reprocuces the setup.

Fixed by ensuring the header height remains stable
during the layout calculation.

Change-Id: I0a4c1228cee0099d2d050490df8b9451072daac4
BCP: 1880686928