[INTERNAL][FIX] fl changes-bundles can be configured to always be loaded
In case the webIde starts an application without a build it is possible
that changes done in the webIde are not correctly fetched in a preview.
To ensure a proper loading of such a bundle - even if not predefined or
in debug mode, a config was introduced to trigger the request for the
changes-bundle.json.
This json still has to be created, as it is done within the webIde
before the preview is started.

Change-Id: I64072ee0e0f19c0116182115851ca2bf4652b606
(cherry picked from commit 3f2a66464035d6ac6601d5935f2c16cc33c74c93)
CR-Id: 002075125900005033182018
BCP: n/a