[FIX] sap.m.Fiori20Adapter: Ensure first NavContainer page is adapted
Background:
Requirement: The HeaderAdapter must adapt any newly displayed page
of the adapted app.

How: The HeaderAdapter listens for the "navigate" event fired for
the newly displayed pages of the NavContainer => adapts the new page

Exceptional case: By design, the "navigate" event is NOT fired for
the *initial* page of the NavContainer
Therefore, for the *initial* page, the HeaderAdapter relies on the
"adaptableContentChange" event that is fired when the initial page
is added.

Problem:
The "adaptableContentChange" event was not fired when the container
was not in dom
=> the HeaderAdapter was not notified of the addition of the page
=> the HeaderAdapter failed to adapt the page

Solution:
Ensured the "adaptableContentChange" is fired whenever an initial
page is added, *even if* the container is not in dom yet

Change-Id: Icd848710bdc53581b6268756d209a62334efe105
BCP: 1870329009
(cherry picked from commit f72e1df5e94319f66ccdcf36a0c5df6f7320e866)
CR-Id: 002075125800002205952020