[FIX] Core: Default to async preload when ui5loader is async
Note that this change only affects the value of the "preload"
configuration, not any behavior.
When using the ui5loader async mode, the initial core bootstrap already
used async loading for the library preloads but did not reflect this in
the preload config that can be retrieved later on.
This mainly caused false positives in the Preload Configuration Support
Rule.

Change-Id: If66dc4750631c8d05b1aa8e90b41aaa5a09aa1f0