[FIX] m.ViewSettingsDialog: Tooltip for sort items is now shown
VSD takes each ViewSettingsItem, which is in the 'sortItems'
aggregation and creates StandardListItem from it. Until now,
only the properties of the original VSI were passed to the new
SLI. Because the tooltip is an aggregation however, we needed
to add additional logic, in order for the SLI to display it.

Change-Id: I78f0b6a3b76a0df30e542da3fa45505aab7a6f61
BCP: 1880607284