[FEATURE] sap.m.Column: enabling sort indicator
A new property has been introduced called the sortIndicator,
which can be used to visually indicate if a column is sorted. Based on
the value of the sortIndicator property, a corresponding sort icon would
be rendered on the column to indicate the sort direction along with its
accessibility attribute.
Resolves:
JIRA: CPOUIFTEAMB-1223

Change-Id: Ic53a9900774e2683f14ccb4135ab4e89f195aaf8