[FIX] m.ViewSettingsDialog: Tooltip for sort items is now shown
VSD takes each ViewSettingsItem, which is in the 'sortItems'
aggregation and creates StandardListItem from it. Until now,
only the properties of the original VSI were passed to the new
SLI. Because the tooltip is an aggregation however, we needed
to add additional logic, in order for the SLI to display it.

Change-Id: I1d5958e64871903521d4b879722d8ce00a425acf
BCP: 1880607284
CR-Id: 002075125800005241932018