[INTERNAL] Core#setThemeRoot: Add forceUpdate parameter
Introduces a new parameter for setThemeRoot to be able to change a theme
root for some libraries of the currently applied theme and trigging an
automatic update of the stylesheets without needing to switch/apply
themes again.
Note that the ThemeChanged event will not be influenced by this and that
parameters consumed via sap.ui.core.Parameters.get() are not forcefully
reloaded from the new location.

CR-Id: 002075125900004585382018
Change-Id: I7b4cb4b7c7da590584facea728de6ddfb32d99b6
BCP: 1880689890
(cherry picked from commit 244f193c66230c71333600f2566c287c511a9501)