[FEATURE] sap.f.DynamicPageTitle: Added snappedTitleOnMobile aggregation
Description:

The only content that is displayed in the DynamicPageTitle
when it is viewed on a phone mobile device and the DynamicPageHeader
is in collapsed (snapped) state.

Using this aggregation enables you to provide a simple, single-line
title that takes less space on the smaller phone screens when the
DynamicPageHeader is in its collapsed (snapped) state.

Change-Id: I111662712617a6b49225c6a9610ecd9e3ce49329
JIRA: BGSOFUIPIRIN-3514