[INTERNAL][FIX] sap.ui.codeeditor: Test suite files are named correctly
Problem: The testsuite was renamed from "testsuite.qunit.html" to
"testsuite.mobilequnit.html", which has the side effect
that ALL tests are no longer taken into account for the
UI5 voter executions, as the trigger project still
references the deleted "testsuite.qunit.html"!

Fix The names are reverted as it was before.

Change-Id: Ic0b861416e28038b73c5973c09a43a2a51a0d73b
BCP: 1870458003