[FIX] sap.m.PlanningCalendar - API appointment (de)selection works now.
Previously if the app. developer used to call Appointment.prototype.setSelected,
an inconsistency in CalendarRow (CalendarRow.aSelectedAppointments)
let to issue when the end-user further selects that appointment.

The resulting issue was was that the end user does not anymore get feedback that the
appointment is selected/deselected.

(cherry picked from commit 9b8e2b1)

Change-Id: I1869e92aa8487e62962845827675d1f01eb7d2ba
BCP: 1880678579
CR-Id: 002075125900003977252018