[FIX] sap.m.NavContainer: JSDoc updated
Previously: We were stating that we can use the "afterShow" event to
focus another element.
This is a misleading information, since if we have "autoFocus" property
enabled, the focus that we set in "afterShow" is lost.

Now: We are stating that we can use "afterShow", only if "autoFocus"
property is set to false.

I double checked this statement and indeed we are losing the focus,
if we use "afterShow" while having "autoFocus" enabled.

Change-Id: I5baef99e05c30bb9a80d2a6a9122cb36e7ef5976
BCP: 1880724353