[FIX] sap.m.Image: Using DetailBox with invalid src results as error
Before: When using DetailBox with invalid src of the Image, no error
is registered.

Now: Error is fired when we have an invalid src for the Image and
DetailBox present.

Root Cause: The $DomNode that we are assigning at onAfterRendering
life-cycle hook was incorrect when we have DetailBox present.

Change-Id: Ifd18b828ca8d5ef568b2702801d4792d6464b92e
BCP: 1870456103