[FIX] sap.m.PlanningCalendar - toolbar content is not cut anymore.
Previously if PlaanningCalendar is rendered in a container whose width
forced (overflow) toolbar to overflow its items, the overflowing did not
work well - toolbar was visually cut and some items were not overflowing.

The reason for that was a wrong internal usage of sap.m.OverflowToolbar,
where explicit overriding of its methods for aggregation 'content' was
not complete.

The fix utilizes the aggregation forwarding core feature, where all
the elements given to the PlanningCalendar's aggregation "toolbarContent"
are being forwarded to the "content" of the "headerContent" of the
internally created sap.m.Table (used for appointments visualization)

BCP: 1870468668
Change-Id: I6c097a6837fb1c7e6c3295580393c16f477355a0