[INTERNAL] core.Element: merge with CustomData to avoid a sync require
The 'official' module for accessing the CustomData class is still
sap/ui/core/CustomData.js. Care has been taken to properly document this
in the API reference.

Additionally, links to the developer guide section about custom data
have been added to both, the CustomData class as well as to the
Element#data method.

Change-Id: I5ff3c135f51caf27b8d5600172609ac5f916a74c